requirements.txt
openai>=1.3.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.0
duckduckgo-search==3.9.3
langchain==0.0.300
langchain-community==0.0.10
chromadb==0.4.18
